Cycles: Add support for building with CUDA 13.0 and OptiX 9.0
The compiler in the CUDA 13 toolkit dropped support for Maxwell, Pascal and
Volta architectures (sm_5X, sm_6X and sm_70), which affects both CUDA and
OptiX kernel compilation for Cycles. This patch makes it so building CUDA
kernel binaries for those architectures are skipped when CUDA 13 is used,
but it will still build them if there is a CUDA 11 toolkit available (e.g.
on buildbot), like how things are handled for other architectures. The
OptiX PTX kernel is compiled with the minimum architecture available
(compute_75 with CUDA 13, compute_50 with previous CUDA versions).

In addition, loading the PTX kernel after initializing OptiX version 9.0
would fail with a OPTIX_ERROR_INVALID_FUNCTION_USE, due to the use of
"optixTrace" within direct callables (as part of the AO and bevel SVM nodes).
Starting with OptiX 9.0 this is no longer allowed, rather one has to use
"optixTraverse" in those cases. This patch thus changes the affected
intersection routines to use "optixTraverse". As a side effect it also
simplifies the `scene_intersect_shadow` function, which no longer invokes
the closest hit program, and can just quickly return hit status. The
minimum OptiX version Cycles requires is already 8.0, which supports
"optixTraverse", so it can just be applied always.

Finally, this patch also adds the `--split-compile=0` argument to nvcc
when available, which tells the compiler to internally split the module
into pieces that can be processed in parallel on multiple threads
(the `=0` notes to use as many threads as there are CPU cores), whic
can greatly improving compile times, while not making compromises on
performance.

Ref #145570

Pull Request: https://projects.blender.org/blender/blender/pulls/145130