Fix #145206: Outliner not using the sequencer scene to show strips
The outliner was still using the active scene to show the sequencer
strips in the VSE mode.

This adds the workspace to the `TreeSourceData` so that
the `TreeDisplaySequencer` can read the sequencer
scene.

Note that more places need to be fixed to make this mode
work correctly again.

Pull Request: https://projects.blender.org/blender/blender/pulls/145247