Fix (unreported) wrong handling of system-defined IDProperties in RNA.
In RNA access code, when trying to 'decode' a given property into valid
data, if the existing storage IDProperty for a runtime-defined RNA
property was detected as invalid (e.g. wrong dimensions for an
array...), code was trying to remove that property from the user-defined
IDProperties (custom properties) instead of the system ones.

Mistake from the original IDProperty split commit (7276b2009a88196).