Fix: Grease Pencil: Pen tool moving incorrect handle when extruding from curve start
Previously, if a user extruded a curve with the Pen tool and held down
the `Ctrl` modifier key, the right handle would always be moved
independently. This should not happen when extruding a curve from
the beginning.

This pull request adds a check that will determine if the left handle
is currently selected when extruding a curve. This is the case, when
extruding a curve from the beginning. Then, the left handle will be
moved independently, instead of the right handle.

Pull Request: https://projects.blender.org/blender/blender/pulls/145235