Fix #143662: Crash with Cycles persistent data and file thumbnail generation
The Embree scene contains a TBB task group that has a parent pointer to the
task group it was created in. In Cycles this task group was only temporarily
created on the stack, resulting in a dangling parent pointer.

The simple solution is to make the Cycles side task group persistent too.

Many thanks to Aras for figuring this one out, this was a very tricky one.

Pull Request: https://projects.blender.org/blender/blender/pulls/145515