Fix: Last row in file browser showing sometimes clipped out of view
In the vertical list display mode, if only the top part of a file was
scrolled into view, the file wouldn't actually draw. So while scrolling
down the list, the last item would visibly pop up. This wasn't intended,
not drawing these items is just an optimization.

Issue was the offset for the table header (reading something like "Name
| Modified | Size") wasn't accounted for.