ffmpeg: cap maximum thread count to 16
ffmpeg internals log a warning when telling it to use more than 16
threads, so cap thread count to that.

Testing the change locally on Ryzen 5950X (where it was using 32
threads previously), capping at 16 is actually faster:
- Playback of one 4K video downsampled at 1080 res:
  seq_render_strip_stack average 9.12ms -> 7.83ms
- Playback of two blender 1080p videos: seq_render_strip_stack
  average 5.79ms -> 4.94ms

I suspect this might be because this CPU has 16 physical cores
with SMT (so 32 "logical" threads), and depending on the workload
capping at number of physical cores might be better.

Pull Request: https://projects.blender.org/blender/blender/pulls/145665