VSE: paste strips at mouse in preview
This PR makes it possible to paste strips directly at the mouse in the
VSE Preview, making it faster to paste strips directly at the position
the user wants the strip to be. It is still possible to paste the strip at the
original position with the `ctrl + shift + v` shortcut.

Pull Request: https://projects.blender.org/blender/blender/pulls/145008