UI: Correction & Improvements to File Browser Font Tooltips
The tooltips for fonts in the file browser, while in list view, now
shows either no preview or a blank white square (depending if you
viewed them with thumbnails in the current session). This PR corrects
that to always show a preview. It also shows this preview while in
thumbnail view.

Pull Request: https://projects.blender.org/blender/blender/pulls/145685