Color Management: Use working space from custom OCIO config in startup blend
When using the default Blender config use what was saved in the startup blend.
But when using a custom config it's likely the scene linear space from that config
is wanted. And not using it can lead to some confusing situations, where you end
up with an ACES config and the Linear Rec.709 working space.

Pull Request: https://projects.blender.org/blender/blender/pulls/145884